refactor: change loader position, edit AddressDto data type, - remove type file in component folder - add set markers and import useMarkers
