+++ src/app/api/type.d.ts
@@ -25,9 +25,9 @@ export interface BasicDto {
 export interface AddressDto {
   gu: SeoulGuType;
 
-  lat: string; //위도
+  lat: number; //위도
 
-  lon: string; //경도
+  lng: number; //경도
 
   fullAddress: string;
 }
